requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.6
aioresponses==0.7.6
aiosignal==1.3.1
asyncio==3.4.3
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
dnspython==2.6.1
frozenlist==1.4.1
httpretty==1.1.4
idna==3.10
lxml==5.3.0
multidict==6.1.0
python-Wappalyzer==0.3.1
requests==2.32.3
setuptools==75.1.0
soupsieve==2.6
urllib3==2.2.3
yarl==1.13.0
